rename ddl/lightning to ddl/ingest
